Add lore and search methods in ItemBuilderMechanic class
This commit introduces new methods that allow users to add, remove, or modify the lore of items and perform a search operation in the ItemBuilderMechanic class. It also adds new getters in the InventoryConfig class and introduces rules for handling null items in the InvCustom class. In addition, the commit includes enhancements to the InvCustomPagesAnvil class such as allowing the update of the search list and managing item interfaces and also adds inventory setting rules for item interfaces.